Create Ajie Felicia ogechukwu 2024704006
This python program Implements a grading system using if and  elif statements. It takes a student's score as input and print the corresponding grade based on the following scales;

-70–100: Grade A
-60–69: Grade B
-50–59: Grade C
-45–49: Grade D
-40–44: Grade E
-0–39: Grade F

For every valid score, the program also print "YES" to confirm the reading was successful. If the score is outside the valid range, it print " invalid score" and "NO".


Created by: Ajie Felicia Ogechukwu
Reg no : 2024704006